Break project license file rule into smaller rules
