+++ src/analyzer.rs
@@ -10,7 +10,9 @@ static RULES: &[&dyn Rule] = &[
   &ProjectImportNamesRule,
   &ProjectNameRule,
   &ProjectDescriptionRule,
-  &ProjectLicenseRule,
+  &ProjectLicenseValueRule,
+  &ProjectLicenseFilesRule,
+  &ProjectLicenseClassifiersRule,
   &ProjectClassifiersRule,
   &ProjectKeywordsRule,
   &ProjectPeopleRule,
