--- src/rule/project_license.rs
@@ -1,803 +1,0 @@
-use super::*;
-
-pub(crate) struct ProjectLicenseRule;
-
-impl Rule for ProjectLicenseRule {
-  fn display_name(&self) -> &'static str {
-    "Project License"
-  }
-
-  fn id(&self) -> &'static str {
-    "project-license"
-  }
-
-  fn run(&self, context: &RuleContext<'_>) -> Vec<lsp::Diagnostic> {
-    if !context.tree().errors.is_empty() {
-      return Vec::new();
-    }
-
-    let Some(project) = context.project() else {
-      return Vec::new();
-    };
-
-    let document = context.document();
-
-    let license = project.try_get("license").ok();
-    let license_files = project.try_get("license-files").ok();
-
-    let classifiers = project.try_get("classifiers").ok();
-
-    let mut diagnostics = Vec::new();
-
-    if let Some(license_files) = &license_files {
-      diagnostics.extend(Self::check_license_files(document, license_files));
-    }
-
-    if let Some(license) = &license {
-      diagnostics.extend(Self::check_license(
-        document,
-        license,
-        license_files.is_some(),
-      ));
-    }
-
-    diagnostics.extend(Self::check_license_classifiers(
-      document,
-      license.as_ref(),
-      classifiers,
-    ));
-
-    diagnostics
-  }
-}
-
-impl ProjectLicenseRule {
-  fn check_license(
-    document: &Document,
-    license: &Node,
-    license_files_present: bool,
-  ) -> Vec<lsp::Diagnostic> {
-    match license {
-      Node::Str(string) => {
-        Self::check_license_string(document, license, string.value())
-      }
-      Node::Table(_) if license_files_present => vec![
-        lsp::Diagnostic {
-          message: "`project.license` must be a string SPDX expression when `project.license-files` is present".to_string(),
-          range: license.range(&document.content),
-          severity: Some(lsp::DiagnosticSeverity::ERROR),
-          ..Default::default()
-        },
-      ],
-      Node::Table(_) => {
-        let mut diagnostics = Vec::new();
-
-        diagnostics.push(lsp::Diagnostic {
-          message:
-            "`project.license` tables are deprecated; prefer a SPDX expression string and `project.license-files`"
-              .to_string(),
-          range: license.range(&document.content),
-          severity: Some(lsp::DiagnosticSeverity::WARNING),
-          ..Default::default()
-        });
-
-        diagnostics.extend(Self::check_table(document, license));
-
-        diagnostics
-      }
-      _ => vec![lsp::Diagnostic {
-        message: "`project.license` must be a string or table".to_string(),
-        range: license.range(&document.content),
-        severity: Some(lsp::DiagnosticSeverity::ERROR),
-        ..Default::default()
-      }],
-    }
-  }
-
-  fn check_license_classifiers(
-    document: &Document,
-    license: Option<&Node>,
-    classifiers: Option<Node>,
-  ) -> Vec<lsp::Diagnostic> {
-    let mut diagnostics = Vec::new();
-
-    let Some(classifiers) = classifiers else {
-      return diagnostics;
-    };
-
-    let Some(array) = classifiers.as_array() else {
-      return diagnostics;
-    };
-
-    let license_is_string = license.is_some_and(Node::is_str);
-
-    let mut has_license_classifier = false;
-
-    for item in array.items().read().iter() {
-      let Some(value) = item.as_str() else {
-        continue;
-      };
-
-      if value.value().starts_with("License ::") {
-        has_license_classifier = true;
-
-        diagnostics.push(lsp::Diagnostic {
-          message: if license_is_string {
-            "`project.classifiers` license classifiers are deprecated when `project.license` is present (use only `project.license`)".to_string()
-          } else {
-            "`project.classifiers` license classifiers are deprecated; use `project.license` instead"
-              .to_string()
-          },
-          range: item.range(&document.content),
-          severity: Some(lsp::DiagnosticSeverity::WARNING),
-          ..Default::default()
-        });
-      }
-    }
-
-    if license_is_string && has_license_classifier {
-      diagnostics.push(lsp::Diagnostic {
-        message:
-          "`project.classifiers` must not include license classifiers when `project.license` is set"
-            .to_string(),
-        range: classifiers.range(&document.content),
-        severity: Some(lsp::DiagnosticSeverity::ERROR),
-        ..Default::default()
-      });
-    }
-
-    diagnostics
-  }
-
-  fn check_license_files(
-    document: &Document,
-    license_files: &Node,
-  ) -> Vec<lsp::Diagnostic> {
-    let mut diagnostics = Vec::new();
-
-    let Some(array) = license_files.as_array() else {
-      diagnostics.push(lsp::Diagnostic {
-        message: "`project.license-files` must be an array of strings"
-          .to_string(),
-        range: license_files.range(&document.content),
-        severity: Some(lsp::DiagnosticSeverity::ERROR),
-        ..Default::default()
-      });
-
-      return diagnostics;
-    };
-
-    let items = array.items().read();
-
-    if items.is_empty() {
-      return diagnostics;
-    }
-
-    let Some(root) = document.root() else {
-      return diagnostics;
-    };
-
-    for item in items.iter() {
-      let Some(pattern) = item.as_str() else {
-        diagnostics.push(lsp::Diagnostic {
-          message: "`project.license-files` items must be strings".to_string(),
-          range: item.range(&document.content),
-          severity: Some(lsp::DiagnosticSeverity::ERROR),
-          ..Default::default()
-        });
-
-        continue;
-      };
-
-      let pattern_value = pattern.value();
-
-      if pattern_value.trim().is_empty() {
-        diagnostics.push(lsp::Diagnostic {
-          message: "`project.license-files` patterns must not be empty"
-            .to_string(),
-          range: item.range(&document.content),
-          severity: Some(lsp::DiagnosticSeverity::ERROR),
-          ..Default::default()
-        });
-
-        continue;
-      }
-
-      if let Err(message) = Self::validate_license_files_pattern(pattern_value)
-      {
-        diagnostics.push(lsp::Diagnostic {
-          message: format!("invalid `project.license-files` pattern `{pattern_value}`: {message}"),
-          range: item.range(&document.content),
-          severity: Some(lsp::DiagnosticSeverity::ERROR),
-          ..Default::default()
-        });
-
-        continue;
-      }
-
-      match Self::matched_files(&root, pattern_value) {
-        Ok(matches) if matches.is_empty() => diagnostics.push(
-          lsp::Diagnostic {
-            message: format!(
-              "`project.license-files` pattern `{pattern_value}` did not match any files"
-            ),
-            range: item.range(&document.content),
-            severity: Some(lsp::DiagnosticSeverity::ERROR),
-            ..Default::default()
-          },
-        ),
-        Ok(matches) => diagnostics.extend(
-          matches
-            .into_iter()
-            .filter_map(|path| Self::ensure_utf8_file(&path).err().map(|message| {
-              lsp::Diagnostic {
-                message,
-                range: item.range(&document.content),
-                severity: Some(lsp::DiagnosticSeverity::ERROR),
-                ..Default::default()
-              }
-            })),
-        ),
-        Err(error) => diagnostics.push(lsp::Diagnostic {
-          message: format!(
-            "failed to evaluate `project.license-files` pattern `{pattern_value}`: {error}"
-          ),
-          range: item.range(&document.content),
-          severity: Some(lsp::DiagnosticSeverity::ERROR),
-          ..Default::default()
-        }),
-      }
-    }
-
-    diagnostics
-  }
-
-  fn check_license_string(
-    document: &Document,
-    license: &Node,
-    value: &str,
-  ) -> Vec<lsp::Diagnostic> {
-    if value.trim().is_empty() {
-      return vec![lsp::Diagnostic {
-        message: "`project.license` must not be empty".to_string(),
-        range: license.range(&document.content),
-        severity: Some(lsp::DiagnosticSeverity::ERROR),
-        ..Default::default()
-      }];
-    }
-
-    let mut diagnostics = Vec::new();
-
-    match spdx::Expression::parse(value) {
-      Ok(expression) => {
-        if let Ok(Some(canonical)) = spdx::Expression::canonicalize(value) {
-          diagnostics.push(lsp::Diagnostic {
-            message: format!(
-              "`project.license` must use a case-normalized SPDX expression (use `{canonical}`)"
-            ),
-            range: license.range(&document.content),
-            severity: Some(lsp::DiagnosticSeverity::ERROR),
-            ..Default::default()
-          });
-        }
-
-        diagnostics.extend(Self::deprecation_warnings(
-          document,
-          license,
-          &expression,
-        ));
-      }
-      Err(error)
-        if matches!(error.reason, spdx::error::Reason::DeprecatedLicenseId) =>
-      {
-        if let Ok(expression) =
-          spdx::Expression::parse_mode(value, spdx::ParseMode::LAX)
-        {
-          diagnostics.extend(Self::deprecation_warnings(
-            document,
-            license,
-            &expression,
-          ));
-        }
-      }
-      Err(error) => {
-        let reason = error.reason.to_string();
-
-        let suggestion = spdx::Expression::canonicalize(value)
-          .ok()
-          .flatten()
-          .map(|canonical| format!(" (did you mean `{canonical}`?)"))
-          .unwrap_or_default();
-
-        diagnostics.push(lsp::Diagnostic {
-          message: format!(
-            "`project.license` must be a valid SPDX expression: {reason}{suggestion}"
-          ),
-          range: license.range(&document.content),
-          severity: Some(lsp::DiagnosticSeverity::ERROR),
-          ..Default::default()
-        });
-      }
-    }
-
-    diagnostics
-  }
-
-  fn check_table(document: &Document, license: &Node) -> Vec<lsp::Diagnostic> {
-    let mut diagnostics = Vec::new();
-
-    let file = license.try_get("file").ok();
-    let text = license.try_get("text").ok();
-
-    match (file.as_ref(), text.as_ref()) {
-      (Some(_), Some(_)) => diagnostics.push(lsp::Diagnostic {
-        message: "`project.license` must specify only one of `file` or `text`"
-          .to_string(),
-        range: license.range(&document.content),
-        severity: Some(lsp::DiagnosticSeverity::ERROR),
-        ..Default::default()
-      }),
-      (None, None) => diagnostics.push(lsp::Diagnostic {
-        message:
-          "missing required key `project.license.file` or `project.license.text`"
-            .to_string(),
-        range: license.range(&document.content),
-        severity: Some(lsp::DiagnosticSeverity::ERROR),
-        ..Default::default()
-      }),
-      _ => {}
-    }
-
-    if let Some(ref file) = file {
-      match file {
-        Node::Str(string) => {
-          diagnostics.extend(
-            document
-              .validate_relative_path(
-                string.value(),
-                "project.license.file",
-                file,
-              )
-              .err()
-              .into_iter()
-              .flatten(),
-          );
-        }
-        _ => diagnostics.push(lsp::Diagnostic {
-          message: "`project.license.file` must be a string".to_string(),
-          range: file.range(&document.content),
-          severity: Some(lsp::DiagnosticSeverity::ERROR),
-          ..Default::default()
-        }),
-      }
-    }
-
-    if let Some(text) = text {
-      match text {
-        Node::Str(_) => {}
-        _ => diagnostics.push(lsp::Diagnostic {
-          message: "`project.license.text` must be a string".to_string(),
-          range: text.range(&document.content),
-          severity: Some(lsp::DiagnosticSeverity::ERROR),
-          ..Default::default()
-        }),
-      }
-    }
-
-    diagnostics
-  }
-
-  fn deprecation_warnings(
-    document: &Document,
-    license: &Node,
-    expression: &spdx::Expression,
-  ) -> Vec<lsp::Diagnostic> {
-    let mut diagnostics = Vec::new();
-
-    let mut seen_licenses = HashSet::new();
-    let mut seen_exceptions = HashSet::new();
-
-    for requirement in expression.requirements() {
-      if let Some(id) = requirement.req.license.id()
-        && id.is_deprecated()
-        && seen_licenses.insert(id.name)
-      {
-        diagnostics.push(lsp::Diagnostic {
-          message: format!(
-            "license identifier `{}` in `project.license` is deprecated",
-            id.name
-          ),
-          range: license.range(&document.content),
-          severity: Some(lsp::DiagnosticSeverity::WARNING),
-          ..Default::default()
-        });
-      }
-
-      if let Some(addition) = &requirement.req.addition
-        && let Some(id) = addition.id()
-        && id.is_deprecated()
-        && seen_exceptions.insert(id.name)
-      {
-        diagnostics.push(lsp::Diagnostic {
-          message: format!(
-            "license exception `{}` in `project.license` is deprecated",
-            id.name
-          ),
-          range: license.range(&document.content),
-          severity: Some(lsp::DiagnosticSeverity::WARNING),
-          ..Default::default()
-        });
-      }
-    }
-
-    diagnostics
-  }
-
-  fn ensure_utf8_file(path: &Path) -> Result<(), String> {
-    fs::read_to_string(path).map(|_| ()).map_err(|error| {
-      format!(
-        "license file `{}` must be valid UTF-8 text ({error})",
-        path.display()
-      )
-    })
-  }
-
-  fn glob_max_depth(pattern: &str) -> Option<usize> {
-    if pattern.contains("**") {
-      return None;
-    }
-
-    Some(
-      pattern
-        .split('/')
-        .filter(|segment| !segment.is_empty())
-        .count()
-        .max(1),
-    )
-  }
-
-  fn matched_files(root: &Path, pattern: &str) -> Result<Vec<PathBuf>, String> {
-    let mut builder =
-      globwalk::GlobWalkerBuilder::from_patterns(root, &[pattern])
-        .follow_links(false);
-
-    if let Some(max_depth) = Self::glob_max_depth(pattern) {
-      // Avoid walking the entire workspace when the pattern does not request
-      // recursive matching.
-      builder = builder.max_depth(max_depth);
-    }
-
-    let walker = builder.build().map_err(|error| error.to_string())?;
-
-    let mut paths = Vec::new();
-
-    for entry in walker {
-      paths.push(entry.map_err(|error| error.to_string())?.into_path());
-    }
-
-    Ok(paths)
-  }
-
-  fn validate_license_files_pattern(pattern: &str) -> Result<(), String> {
-    if pattern.starts_with('/') {
-      return Err(
-        "patterns must be relative; leading `/` is not allowed".into(),
-      );
-    }
-
-    if pattern.contains('\\') {
-      return Err("path delimiter must be `/`, not `\\`".into());
-    }
-
-    if pattern.contains("..") {
-      return Err("parent directory segments (`..`) are not allowed".into());
-    }
-
-    let mut in_brackets = false;
-
-    for (index, character) in pattern.chars().enumerate() {
-      match character {
-        '[' if !in_brackets => in_brackets = true,
-        ']' if in_brackets => in_brackets = false,
-        ']' => {
-          return Err(format!(
-            "unmatched closing `]` at position {}",
-            index + 1
-          ));
-        }
-        _ if in_brackets => {
-          if !matches!(character, 'A'..='Z' | 'a'..='z' | '0'..='9' | '_' | '-' | '.')
-          {
-            return Err("`[]` character classes may only contain alphanumerics, `_`, `-`, or `.`"
-              .into());
-          }
-        }
-        '*' | '?' | '/' | '.' | '-' | '_' => {}
-        c if c.is_ascii_alphanumeric() => {}
-        c => {
-          return Err(format!("character `{c}` is not allowed"));
-        }
-      }
-    }
-
-    if in_brackets {
-      return Err("unclosed `[` in pattern".into());
-    }
-
-    Ok(())
-  }
-}
-
-#[cfg(test)]
-mod tests {
-  use {super::*, pretty_assertions::assert_eq};
-
-  #[test]
-  fn validate_license_files_pattern_valid_simple() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("LICENSE"),
-      Ok(())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_valid_with_extension() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("LICENSE.txt"),
-      Ok(())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_valid_with_subdirectory() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("licenses/MIT.txt"),
-      Ok(())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_valid_with_wildcard() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("LICENSE*"),
-      Ok(())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_valid_with_question_mark() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("LICENSE?"),
-      Ok(())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_valid_with_globstar() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("**/LICENSE"),
-      Ok(())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_valid_with_character_class() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("LICENSE[0-9].txt"),
-      Ok(())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_valid_with_underscore_dash() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("LICENSE_MIT-2.0.txt"),
-      Ok(())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_valid_complex() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern(
-        "licenses/**/LICENSE*.txt"
-      ),
-      Ok(())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_valid_character_class_alphanumeric() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("file[abc123].txt"),
-      Ok(())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_valid_character_class_with_underscore() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("file[a_b].txt"),
-      Ok(())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_valid_character_class_with_dash() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("file[a-z].txt"),
-      Ok(())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_valid_character_class_with_dot() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("file[a.b].txt"),
-      Ok(())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_leading_slash() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("/LICENSE"),
-      Err("patterns must be relative; leading `/` is not allowed".to_string())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_backslash() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("licenses\\LICENSE"),
-      Err("path delimiter must be `/`, not `\\`".to_string())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_parent_directory() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("../LICENSE"),
-      Err("parent directory segments (`..`) are not allowed".to_string())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_parent_directory_in_middle() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("foo/../LICENSE"),
-      Err("parent directory segments (`..`) are not allowed".to_string())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_unmatched_closing_bracket() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("LICENSE].txt"),
-      Err("unmatched closing `]` at position 8".to_string())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_unclosed_bracket() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("LICENSE[abc"),
-      Err("unclosed `[` in pattern".to_string())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_invalid_character_in_class() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("LICENSE[a/b].txt"),
-      Err("`[]` character classes may only contain alphanumerics, `_`, `-`, or `.`".to_string())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_invalid_character_space() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("LICENSE FILE.txt"),
-      Err("character ` ` is not allowed".to_string())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_invalid_character_special() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("LICENSE@.txt"),
-      Err("character `@` is not allowed".to_string())
-    );
-  }
-
-  #[test]
-  fn validate_license_files_pattern_invalid_character_hash() {
-    assert_eq!(
-      ProjectLicenseRule::validate_license_files_pattern("LICENSE#1.txt"),
-      Err("character `#` is not allowed".to_string())
-    );
-  }
-
-  #[test]
-  fn glob_max_depth_simple_file() {
-    assert_eq!(ProjectLicenseRule::glob_max_depth("LICENSE"), Some(1));
-  }
-
-  #[test]
-  fn glob_max_depth_with_extension() {
-    assert_eq!(ProjectLicenseRule::glob_max_depth("LICENSE.txt"), Some(1));
-  }
-
-  #[test]
-  fn glob_max_depth_one_subdirectory() {
-    assert_eq!(
-      ProjectLicenseRule::glob_max_depth("licenses/LICENSE"),
-      Some(2)
-    );
-  }
-
-  #[test]
-  fn glob_max_depth_two_subdirectories() {
-    assert_eq!(
-      ProjectLicenseRule::glob_max_depth("foo/bar/LICENSE"),
-      Some(3)
-    );
-  }
-
-  #[test]
-  fn glob_max_depth_with_trailing_slash() {
-    assert_eq!(ProjectLicenseRule::glob_max_depth("licenses/"), Some(1));
-  }
-
-  #[test]
-  fn glob_max_depth_with_leading_slash() {
-    assert_eq!(
-      ProjectLicenseRule::glob_max_depth("/licenses/LICENSE"),
-      Some(2)
-    );
-  }
-
-  #[test]
-  fn glob_max_depth_with_wildcard() {
-    assert_eq!(
-      ProjectLicenseRule::glob_max_depth("licenses/*.txt"),
-      Some(2)
-    );
-  }
-
-  #[test]
-  fn glob_max_depth_with_globstar() {
-    assert_eq!(ProjectLicenseRule::glob_max_depth("**/LICENSE"), None);
-  }
-
-  #[test]
-  fn glob_max_depth_globstar_in_middle() {
-    assert_eq!(
-      ProjectLicenseRule::glob_max_depth("licenses/**/LICENSE"),
-      None
-    );
-  }
-
-  #[test]
-  fn glob_max_depth_globstar_at_end() {
-    assert_eq!(ProjectLicenseRule::glob_max_depth("licenses/**"), None);
-  }
-
-  #[test]
-  fn glob_max_depth_multiple_consecutive_slashes() {
-    assert_eq!(
-      ProjectLicenseRule::glob_max_depth("foo//bar///LICENSE"),
-      Some(3)
-    );
-  }
-
-  #[test]
-  fn glob_max_depth_empty_string() {
-    assert_eq!(ProjectLicenseRule::glob_max_depth(""), Some(1));
-  }
-
-  #[test]
-  fn glob_max_depth_single_slash() {
-    assert_eq!(ProjectLicenseRule::glob_max_depth("/"), Some(1));
-  }
-}
